confirming that the value received by huffman function is a double type

+++ functions/sourceDecoding.m
@@ -6,7 +6,7 @@ function [ sig ] = sourceDecoding( comp_sig )
     
     [syms, p] = imagesHuffman('images'); % Defining huffman requirements
     dict = huffmandict(syms, p); % Huffman Dictionary
-    sig = huffmandeco(comp_sig, dict); % Decoding the signal
+    sig = huffmandeco(double(comp_sig), dict); % Decoding the signal
     sig = reshape(sig, [256,256]); % Reshape signal to image dimensions
     sig = uint8(idct2(single(sig)));  % to image domain
